local/visual-web-editor/ai-agent/requirements.txt
# Visual Web Editor AI Agent Dependencies

# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0

# AI/ML Libraries
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.8.2

# Natural Language Processing
textblob==0.17.1
gensim==4.3.2

# Code Analysis
ast-tools==0.1.0
black==23.11.0
autopep8==2.0.4

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Caching and Performance
redis==5.0.1
joblib==1.3.2
memory-profiler==0.61.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Development Tools
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Utilities
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Optional GPU Support (uncomment if using CUDA)
# torch-audio==2.1.1
# torchvision==0.16.1

# Development Dependencies
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6
